Use uint16_t for our various header sizes in bytes (#2836)
PacketBuffer uses uint16_t for all of its sizes already.  It's easier
to tell that we're not overflowing the buffer if we know our header
sizes also fit in uint16_t.

+++ src/transport/SecurePairingSession.cpp
@@ -116,8 +116,8 @@ CHIP_ERROR SecurePairingSession::AttachHeaderAndSend(uint8_t msgType, System::Pa
         .SetMessageType(msgType) //
         .SetProtocolID(kSecurePairingProtocol);
 
-    size_t headerSize              = payloadHeader.EncodeSizeBytes();
-    size_t actualEncodedHeaderSize = 0;
+    size_t headerSize                = payloadHeader.EncodeSizeBytes();
+    uint16_t actualEncodedHeaderSize = 0;
 
     VerifyOrExit(msgBuf->EnsureReservedSize(headerSize), err = CHIP_ERROR_NO_MEMORY);
 
@@ -363,8 +363,8 @@ exit:
 
 CHIP_ERROR SecurePairingSession::HandlePeerMessage(const PacketHeader & packetHeader, System::PacketBuffer * msg)
 {
-    CHIP_ERROR err    = CHIP_NO_ERROR;
-    size_t headerSize = 0;
+    CHIP_ERROR err      = CHIP_NO_ERROR;
+    uint16_t headerSize = 0;
     PayloadHeader payloadHeader;
 
     VerifyOrExit(msg != nullptr, err = CHIP_ERROR_INVALID_ARGUMENT);
